log facility overhaul: remove logf() and add date/time support.
logf() is obsolete now, the compiler now picks the correct log function
itself: if only one parameter is given, this is assumed to be a literal
string and it's copied to the log verbatim. If more parameters are
given, the first is a format string and it's handed to vsnprintf
together with the other parameters.

Depending on two new flags, a date and or time stand be prepended to log
messages (in the log buffer).

+++ dispatch.c
@@ -130,7 +130,7 @@ static void generic_task_handler(unsigned int prio, task_id_t command, unsigned 
 	if(stat_task_current_queue[prio] > 0)
 		stat_task_current_queue[prio]--;
 	else
-		logf("task queue %u underrun\n", prio);
+		log("task queue %u underrun\n", prio);
 
 	switch(command)
 	{
@@ -397,7 +397,7 @@ static void slow_timer_callback(void *arg)
 			command_left_to_read.timeout--;
 		else
 		{
-			logf("dispatch: static length command timeout at %u bytes, cleared command buffer\n", command_left_to_read.length);
+			log("dispatch: static length command timeout at %u bytes, cleared command buffer\n", command_left_to_read.length);
 			command_left_to_read.length = 0;
 			command_left_to_read.timeout = 0;
 			string_clear(&command_socket_receive_buffer);
