+++ i2c_sensor.c
@@ -1353,7 +1353,7 @@ static i2c_error_t si114x_sendcmd(si114x_command_t command, unsigned int *respon
 
 		if(local_response != 0x00)
 		{
-			logf("si114x: response not 0: %u\n", local_response);
+			log("si114x: response not 0: %u\n", local_response);
 			goto failed;
 		}
 
@@ -1375,12 +1375,12 @@ static i2c_error_t si114x_sendcmd(si114x_command_t command, unsigned int *respon
 
 			if((local_response & 0b10001000) == 0b10001000)		// overflow, treat as OK
 			{
-				logf("si114x: response overflow: %x\n", local_response);
+				log("si114x: response overflow: %x\n", local_response);
 				break;
 			}
 
-			logf("si114x: response invalid command / timeout: %x\n", local_response);
-			logf("si114x: attempt2: %u\n", attempt2);
+			log("si114x: response invalid command / timeout: %x\n", local_response);
+			log("si114x: attempt2: %u\n", attempt2);
 
 			msleep(1);
 		}
@@ -1389,7 +1389,7 @@ static i2c_error_t si114x_sendcmd(si114x_command_t command, unsigned int *respon
 			break;
 
 failed:
-		logf("si114x: attempt1: %u\n", attempt1);
+		log("si114x: attempt1: %u\n", attempt1);
 		msleep(1);
 	}
 
@@ -3647,7 +3647,7 @@ static i2c_error_t sensor_ccs811_read(i2c_sensor_data_t *data, i2c_sensor_value_
 			return(error);
 		}
 
-		logf("ccs811: error %x\n", i2c_buffer[0]);
+		log("ccs811: error %x\n", i2c_buffer[0]);
 
 		return(i2c_error_device_error_1);
 	}
@@ -7660,7 +7660,7 @@ static i2c_error_t i2c_sensors_detect(void)
 	if(data_entry->basic.id != sensor_info.detect_current_sensor)
 	{
 		sensor_info.detect_failed++;
-		logf("i2c sensor detect: sensor id != index: %u, %u; %u\n", data_entry->basic.id, sensor_info.detect_current_sensor, sizeof(data_entry->basic));
+		log("i2c sensor detect: sensor id != index: %u, %u; %u\n", data_entry->basic.id, sensor_info.detect_current_sensor, sizeof(data_entry->basic));
 		goto abort;
 	}
